refactor metro init and some code restructuring
