+++ cmd/service/metro/metro.go
@@ -2,12 +2,16 @@ package metro
 
 import (
 	"context"
+	"log"
+	"os"
+	"os/signal"
+	"syscall"
 
+	"github.com/razorpay/metro/pkg/logger"
+
+	"github.com/razorpay/metro/internal/boot"
 	"github.com/razorpay/metro/internal/config"
-	"github.com/razorpay/metro/service"
-	"github.com/razorpay/metro/service/producer"
-	pull_consumer "github.com/razorpay/metro/service/pull-consumer"
-	push_consumer "github.com/razorpay/metro/service/push-consumer"
+	config_reader "github.com/razorpay/metro/pkg/config"
 )
 
 const (
@@ -20,10 +24,11 @@ const (
 )
 
 var validComponents = []string{Producer, PullConsumer, PushConsumer}
+var component *Component
 
-// IsValidComponent validates if the input comonent is a valid metro component
+// isValidComponent validates if the input component is a valid metro component
 // validComponents : producer, pull-consumer, push-consumer
-func IsValidComponent(component string) bool {
+func isValidComponent(component string) bool {
 	for _, s := range validComponents {
 		if s == component {
 			return true
@@ -32,46 +37,71 @@ func IsValidComponent(component string) bool {
 	return false
 }
 
-// Component is a holder for a metro's deployable component
-type Component struct {
-	name    string
-	cfg     *config.ComponentConfig
-	service service.IService
-}
-
-// NewComponent returns a new instance of a metro service component
-func NewComponent(component string, cfg *config.ComponentConfig) (*Component, error) {
-	return &Component{
-		cfg:  cfg,
-		name: component,
-	}, nil
-}
-
-// Start a metro component
-func (c *Component) Start(ctx context.Context) <-chan error {
-	errChan := make(chan error)
-	c.service = c.start(ctx, errChan)
-	return errChan
-}
-
-// Stop a metro component
-func (c *Component) Stop() error {
-	return c.service.Stop()
-}
-
-func (c *Component) start(ctx context.Context, errChan chan<- error) service.IService {
-	var svc service.IService
-
-	switch c.name {
-	case Producer:
-		svc = producer.NewService(ctx, c.cfg)
-	case PushConsumer:
-		svc = push_consumer.NewService(ctx, c.cfg)
-	case PullConsumer:
-		svc = pull_consumer.NewService(ctx, c.cfg)
+// Init initializes all modules (logger, tracing, config, metro component)
+func Init(env string, componentName string) {
+	// componentName validation
+	ok := isValidComponent(componentName)
+	if !ok {
+		log.Fatalf("invalid componentName name input : %v", componentName)
 	}
 
-	go svc.Start(errChan)
+	// read the componentName config for env
+	var appConfig config.Config
+	err := config_reader.NewDefaultConfig().Load(env, &appConfig)
+	if err != nil {
+		log.Fatal(err)
+	}
 
-	return svc
+	componentConfig, ok := appConfig[componentName]
+
+	if !ok {
+		log.Fatal("%v config missing", componentName)
+	}
+
+	err = boot.InitMonitoring(env, &componentConfig)
+
+	if err != nil {
+		log.Fatal("error in setting up monitoring : %v", err)
+	}
+
+	// Init the requested componentName
+	component, err = NewComponent(componentName, &componentConfig)
+	if err != nil {
+		log.Fatal("error in creating metro component : %v", err)
+	}
+}
+
+// Run handles the component execution lifecycle
+func Run(ctx context.Context) {
+	// Shutdown tracer
+	defer func() {
+		err := boot.Closer.Close()
+		if err != nil {
+			log.Fatalf("error closing tracer: %v", err)
+		}
+	}()
+
+	errChan := component.Start(ctx)
+
+	// Handle SIGINT & SIGTERM - Shutdown gracefully
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, syscall.SIGTERM, syscall.SIGINT)
+
+	// Block until signal is received or error is received in starting the component.
+	select {
+	case <-c:
+		logger.Ctx(ctx).Infow("received signal")
+	case err := <-errChan:
+		logger.Ctx(ctx).Fatalw("error in starting component", "msg", err.Error())
+	}
+
+	logger.Ctx(ctx).Infow("stopping metro")
+
+	// stop component
+	err := component.Stop()
+	if err != nil {
+		logger.Ctx(ctx).Fatalw("error in stopping metro")
+	}
+
+	logger.Ctx(ctx).Infow("stopped metro")
 }
