+++ internal/boot/boot.go
@@ -2,23 +2,17 @@ package boot
 
 import (
 	"context"
-	"fmt"
 	"io"
-	"log"
 	"os"
 
 	"github.com/opentracing/opentracing-go"
 	"github.com/razorpay/metro/internal/config"
-	config_reader "github.com/razorpay/metro/pkg/config"
 	logpkg "github.com/razorpay/metro/pkg/logger"
 	sentrypkg "github.com/razorpay/metro/pkg/monitoring/sentry"
 	"github.com/razorpay/metro/pkg/tracing"
 )
 
 var (
-	// ComponentConfig contains component configuration values.
-	ComponentConfig config.ComponentConfig
-
 	// Tracer is used for creating spans for distributed tracing
 	Tracer opentracing.Tracer
 	// Closer holds an instance to the RequestTracing object's Closer.
@@ -36,50 +30,24 @@ func GetEnv() string {
 	return environment
 }
 
-// initialize all core dependencies for the application
-func initialize(ctx context.Context, env string, component string) error {
-	// read the component config for env
-	var appConfig config.Config
-	err := config_reader.NewDefaultConfig().Load(env, &appConfig)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	componentConfig, ok := appConfig[component]
-
-	if !ok {
-		log.Fatal(fmt.Errorf("`%v` component missing config", component))
-	}
-
-	ComponentConfig = componentConfig
-
+// InitMonitoring is used to setup logger, tracing and sentry for monitoring
+func InitMonitoring(env string, config *config.ComponentConfig) error {
 	// Initializes Sentry monitoring client.
-	sentry, err := sentrypkg.InitSentry(ComponentConfig.Sentry, env)
+	sentry, err := sentrypkg.InitSentry(config.Sentry, env)
 	if err != nil {
 		return err
 	}
 	// Initializes logging driver.
 	servicekv := map[string]interface{}{
-		"appEnv":        ComponentConfig.App.Env,
-		"serviceName":   ComponentConfig.App.ServiceName,
-		"gitCommitHash": ComponentConfig.App.GitCommitHash,
+		"appEnv":        config.App.Env,
+		"serviceName":   config.App.ServiceName,
+		"gitCommitHash": config.App.GitCommitHash,
 	}
 	logger, err := logpkg.NewLogger(env, servicekv, sentry)
 	if err != nil {
 		return err
 	}
-	Tracer, Closer, err = tracing.Init(ComponentConfig.Tracing, logger.Desugar())
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// InitMetro initializes all packages (logger, tracing, config, metro component)
-func InitMetro(ctx context.Context, env string, service string) error {
-	err := initialize(ctx, env, service)
-
+	Tracer, Closer, err = tracing.Init(config.Tracing, logger.Desugar())
 	if err != nil {
 		return err
 	}
@@ -93,9 +61,6 @@ func NewContext(ctx context.Context) context.Context {
 	if ctx == nil {
 		ctx = context.Background()
 	}
-	//for k, v := range structs.Map(Config.Core) {
-	//  key := strings.ToLower(k)
-	//	ctx = context.WithValue(ctx, key, v)
-	//}
+
 	return ctx
 }
